requirements.txt
python-frontmatter>=1.1.0
pyyaml>=6.0
pytest>=8.0
mkdocs-material>=9.0
